[config_builder_unexport] move balancer config package into cluster_resolver and unexport
