+++ xds/internal/balancer/clusterresolver/configbuilder.go
@@ -16,17 +16,7 @@
  *
  */
 
-// Package balancerconfig contains utility functions to build balancer config.
-// The built config will generate a tree of balancers with priority,
-// cluster_impl, weighted_target, lrs, and roundrobin.
-//
-// This is in a subpackage of cluster_resolver so that it can be used by the EDS
-// balancer. Eventually we will delete the EDS balancer, and replace it with
-// cluster_resolver, then we can move the functions to package cluster_resolver,
-// and unexport them.
-//
-// TODO: move and unexport. Read above.
-package balancerconfig
+package clusterresolver
 
 import (
 	"encoding/json"
@@ -47,22 +37,22 @@ import (
 
 const million = 1000000
 
-// PriorityConfig is config for one priority. For example, if there an EDS and a
-// DNS, the priority list will be [priorityConfig{EDS}, PriorityConfig{DNS}].
+// priorityConfig is config for one priority. For example, if there an EDS and a
+// DNS, the priority list will be [priorityConfig{EDS}, priorityConfig{DNS}].
 //
-// Each PriorityConfig corresponds to one discovery mechanism from the LBConfig
+// Each priorityConfig corresponds to one discovery mechanism from the LBConfig
 // generated by the CDS balancer. The CDS balancer resolves the cluster name to
 // an ordered list of discovery mechanisms (if the top cluster is an aggregated
 // cluster), one for each underlying cluster.
-type PriorityConfig struct {
-	Mechanism DiscoveryMechanism
-	// EDSResp is set only if type is EDS.
-	EDSResp xdsclient.EndpointsUpdate
-	// Addresses is set only if type is DNS.
-	Addresses []string
+type priorityConfig struct {
+	mechanism DiscoveryMechanism
+	// edsResp is set only if type is EDS.
+	edsResp xdsclient.EndpointsUpdate
+	// addresses is set only if type is DNS.
+	addresses []string
 }
 
-// BuildPriorityConfigJSON builds balancer config for the passed in
+// buildPriorityConfigJSON builds balancer config for the passed in
 // priorities.
 //
 // The built tree of balancers (see test for the output struct).
@@ -94,7 +84,7 @@ type PriorityConfig struct {
 //
 // TODO: support setting locality picking policy, and add a parameter for
 // locality picking policy.
-func BuildPriorityConfigJSON(priorities []PriorityConfig, endpointPickingPolicy *internalserviceconfig.BalancerConfig) ([]byte, []resolver.Address, error) {
+func buildPriorityConfigJSON(priorities []priorityConfig, endpointPickingPolicy *internalserviceconfig.BalancerConfig) ([]byte, []resolver.Address, error) {
 	pc, addrs := buildPriorityConfig(priorities, endpointPickingPolicy)
 	ret, err := json.Marshal(pc)
 	if err != nil {
@@ -103,15 +93,15 @@ func BuildPriorityConfigJSON(priorities []PriorityConfig, endpointPickingPolicy 
 	return ret, addrs, nil
 }
 
-func buildPriorityConfig(priorities []PriorityConfig, endpointPickingPolicy *internalserviceconfig.BalancerConfig) (*priority.LBConfig, []resolver.Address) {
+func buildPriorityConfig(priorities []priorityConfig, endpointPickingPolicy *internalserviceconfig.BalancerConfig) (*priority.LBConfig, []resolver.Address) {
 	var (
 		retConfig = &priority.LBConfig{Children: make(map[string]*priority.Child)}
 		retAddrs  []resolver.Address
 	)
 	for i, p := range priorities {
-		switch p.Mechanism.Type {
+		switch p.mechanism.Type {
 		case DiscoveryMechanismTypeEDS:
-			names, configs, addrs := buildClusterImplConfigForEDS(i, p.EDSResp, p.Mechanism, endpointPickingPolicy)
+			names, configs, addrs := buildClusterImplConfigForEDS(i, p.edsResp, p.mechanism, endpointPickingPolicy)
 			retConfig.Priorities = append(retConfig.Priorities, names...)
 			for n, c := range configs {
 				retConfig.Children[n] = &priority.Child{
@@ -122,7 +112,7 @@ func buildPriorityConfig(priorities []PriorityConfig, endpointPickingPolicy *int
 			}
 			retAddrs = append(retAddrs, addrs...)
 		case DiscoveryMechanismTypeLogicalDNS:
-			name, config, addrs := buildClusterImplConfigForDNS(i, p.Addresses)
+			name, config, addrs := buildClusterImplConfigForDNS(i, p.addresses)
 			retConfig.Priorities = append(retConfig.Priorities, name)
 			retConfig.Children[name] = &priority.Child{
 				Config: &internalserviceconfig.BalancerConfig{Name: clusterimpl.Name, Config: config},
