+++ xds/internal/balancer/clusterresolver/configbuilder_test.go
@@ -1,3 +1,5 @@
+// +build go1.12
+
 /*
  *
  * Copyright 2021 gRPC authors.
@@ -16,7 +18,7 @@
  *
  */
 
-package balancerconfig
+package clusterresolver
 
 import (
 	"bytes"
@@ -41,10 +43,8 @@ import (
 )
 
 const (
-	testClusterName     = "test-cluster-name"
 	testLRSServer       = "test-lrs-server"
 	testMaxRequests     = 314
-	testEDSServcie      = "test-eds-service-name"
 	testEDSServiceName  = "service-name-from-parent"
 	testDropCategory    = "test-drops"
 	testDropOverMillion = 1
@@ -123,16 +123,16 @@ func init() {
 // TestBuildPriorityConfigJSON is a sanity check that the built balancer config
 // can be parsed. The behavior test is covered by TestBuildPriorityConfig.
 func TestBuildPriorityConfigJSON(t *testing.T) {
-	gotConfig, _, err := BuildPriorityConfigJSON([]PriorityConfig{
+	gotConfig, _, err := buildPriorityConfigJSON([]priorityConfig{
 		{
-			Mechanism: DiscoveryMechanism{
+			mechanism: DiscoveryMechanism{
 				Cluster:                 testClusterName,
 				LoadReportingServerName: newString(testLRSServer),
 				MaxConcurrentRequests:   newUint32(testMaxRequests),
 				Type:                    DiscoveryMechanismTypeEDS,
 				EDSServiceName:          testEDSServiceName,
 			},
-			EDSResp: xdsclient.EndpointsUpdate{
+			edsResp: xdsclient.EndpointsUpdate{
 				Drops: []xdsclient.OverloadDropConfig{
 					{
 						Category:    testDropCategory,
@@ -149,10 +149,10 @@ func TestBuildPriorityConfigJSON(t *testing.T) {
 			},
 		},
 		{
-			Mechanism: DiscoveryMechanism{
+			mechanism: DiscoveryMechanism{
 				Type: DiscoveryMechanismTypeLogicalDNS,
 			},
-			Addresses: testAddressStrs[4],
+			addresses: testAddressStrs[4],
 		},
 	}, nil)
 	if err != nil {
@@ -173,16 +173,16 @@ func TestBuildPriorityConfigJSON(t *testing.T) {
 }
 
 func TestBuildPriorityConfig(t *testing.T) {
-	gotConfig, gotAddrs := buildPriorityConfig([]PriorityConfig{
+	gotConfig, gotAddrs := buildPriorityConfig([]priorityConfig{
 		{
-			Mechanism: DiscoveryMechanism{
+			mechanism: DiscoveryMechanism{
 				Cluster:                 testClusterName,
 				LoadReportingServerName: newString(testLRSServer),
 				MaxConcurrentRequests:   newUint32(testMaxRequests),
 				Type:                    DiscoveryMechanismTypeEDS,
 				EDSServiceName:          testEDSServiceName,
 			},
-			EDSResp: xdsclient.EndpointsUpdate{
+			edsResp: xdsclient.EndpointsUpdate{
 				Drops: []xdsclient.OverloadDropConfig{
 					{
 						Category:    testDropCategory,
@@ -199,10 +199,10 @@ func TestBuildPriorityConfig(t *testing.T) {
 			},
 		},
 		{
-			Mechanism: DiscoveryMechanism{
+			mechanism: DiscoveryMechanism{
 				Type: DiscoveryMechanismTypeLogicalDNS,
 			},
-			Addresses: testAddressStrs[4],
+			addresses: testAddressStrs[4],
 		},
 	}, nil)
 
@@ -721,14 +721,6 @@ func TestLocalitiesToWeightedTarget(t *testing.T) {
 			}
 		})
 	}
-}
-
-func newString(s string) *string {
-	return &s
-}
-
-func newUint32(i uint32) *uint32 {
-	return &i
 }
 
 func assertString(f func() (string, error)) string {
